Reintroduce "1492 refactor manuscript upload" (#1531)
This reverts commit 10287931b763de7bf3ca7a39162363f2e072d4bb.
