+++ server/xpub-controller/helpers/index.js
@@ -1,0 +1,2 @@
+exports.FilesHelper = require('./files')
+exports.ManuscriptHelper = require('./manuscript')
